lc-gmails/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.111.0
fastapi-cli==0.0.8
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.8
langchain-openai==0.1.7
langchain-text-splitters==0.2.1
langsmith==0.1.81
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.26.4
openai==1.55.3
openpyxl==3.1.2
orjson==3.11.0
packaging==24.2
pandas==2.2.2
propcache==0.3.2
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
#"psycopg[binary]==3.2.1"
sse-starlette==0.10.0
itsdangerous==2.2.0
google-auth-oauthlib==1.2.2
google-api-python-client==2.176.0